pem require optimizations/fixes
