+++ pem/lib/pem.rb
@@ -1,21 +1,4 @@
-require 'pem/manager'
-require 'pem/options'
+require_relative 'pem/manager'
+require_relative 'pem/options'
 
-require 'fastlane_core'
-
-module PEM
-  # Use this to just setup the configuration attribute and set it later somewhere else
-  class << self
-    attr_accessor :config
-  end
-
-  TMP_FOLDER = "/tmp/PEM/"
-  FileUtils.mkdir_p TMP_FOLDER
-
-  ENV['FASTLANE_TEAM_ID'] ||= ENV["PEM_TEAM_ID"]
-  ENV['DELIVER_USER'] ||= ENV["PEM_USERNAME"]
-
-  Helper = FastlaneCore::Helper # you gotta love Ruby: Helper.* should use the Helper class contained in FastlaneCore
-  UI = FastlaneCore::UI
-  ROOT = Pathname.new(File.expand_path('../..', __FILE__))
-end
+require_relative 'pem/module'
