+++ pem/lib/pem/commands_generator.rb
@@ -1,5 +1,9 @@
 require 'commander'
+
 require 'fastlane/version'
+require 'fastlane_core/configuration/configuration'
+require_relative 'options'
+require_relative 'manager'
 
 HighLine.track_eof = false
